api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.37
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.4
